+++ src/dymad/transform/base.py
